+++ feature_tests/environment.py
@@ -1,8 +1,8 @@
 from behave.model import Scenario
-from behave.runner import Context
 
 from feature_tests.common.config_setup import config_setup
-from feature_tests.common.models import TestConfig
+from feature_tests.common.constants import TestMode
+from feature_tests.common.models import Context, TestConfig
 
 
 def before_all(context: Context):
@@ -12,6 +12,11 @@ def before_all(context: Context):
 
 def before_scenario(context: Context, scenario: Scenario):
     context.test_config = config_setup(context=context, scenario_name=scenario.name)
+    if (
+        "integration-only" in scenario.tags
+        and context.test_config.mode is TestMode.LOCAL_TEST
+    ):
+        scenario.mark_skipped()
 
 
 def after_scenario(context: Context, scenario: Scenario):